refactor: moved common service to component service for modularisation
