+++ FE/angular/src/app/components/movie/movie.module.ts
@@ -5,9 +5,11 @@ import { MovieRoutingModule } from './movie-routing.module';
 import { MovieComponent } from './movie.component';
 import { GoBackModule } from '../navigation/go-back/go-back.module';
 import { GoImdbModule } from '../navigation/go-imdb/go-imdb.module';
+import { MovieDataService } from './data.service';
 
 @NgModule({
   declarations: [MovieComponent],
-  imports: [CommonModule, MovieRoutingModule, GoBackModule, GoImdbModule]
+  imports: [CommonModule, MovieRoutingModule, GoBackModule, GoImdbModule],
+  providers: [MovieDataService]
 })
 export class MovieModule {}
