+++ FE/angular/src/app/components/movies/movies.models.ts
@@ -1,0 +1,12 @@
+import { Movie, MovieComplete } from '../movie/movie.models';
+
+export interface SearchResults {
+  Response: string;
+  Search: Movie[];
+  totalResults: string;
+}
+
+export interface MovieData {
+  Decades: number[];
+  Search: MovieComplete[];
+}
