+++ FE/angular/src/app/constants/constant.ts
@@ -1,0 +1,5 @@
+export class Constants {
+  static posterUrl = 'https://m.media-amazon.com/images/M/';
+  static replacePosterUrl = '/assets/images/';
+  static serviceUrl = 'https://www.omdbapi.com/?apikey=f59b2e4b&';
+}
